requirements.txt
Django==2.2.6
django-tempus-dominus==5.1.2.8
pytz==2019.2
sqlparse==0.3.0
